+++ operator/src/main/java/oracle/kubernetes/operator/webhooks/WebhookRestServer.java
@@ -20,7 +20,7 @@ import oracle.kubernetes.operator.http.rest.ExceptionMapper;
 import oracle.kubernetes.operator.http.rest.RequestDebugLoggingFilter;
 import oracle.kubernetes.operator.http.rest.ResponseDebugLoggingFilter;
 import oracle.kubernetes.operator.http.rest.RestConfig;
-import oracle.kubernetes.operator.webhooks.resource.AdmissionWebhookResource;
+import oracle.kubernetes.operator.webhooks.resource.AdmissionWebhookDomainResource;
 import oracle.kubernetes.operator.webhooks.resource.ConversionWebhookResource;
 import oracle.kubernetes.operator.work.Container;
 import org.glassfish.grizzly.http.server.HttpServer;
@@ -85,7 +85,8 @@ public class WebhookRestServer extends BaseRestServer {
         .register(RequestDebugLoggingFilter.class)
         .register(ResponseDebugLoggingFilter.class)
         .register(ExceptionMapper.class)
-        .packages(ConversionWebhookResource.class.getPackageName(), AdmissionWebhookResource.class.getPackageName())
+        .packages(ConversionWebhookResource.class.getPackageName(),
+            AdmissionWebhookDomainResource.class.getPackageName())
         .setProperties(Map.of(RestConfig.REST_CONFIG_PROPERTY, restConfig));
   }
