+++ operator/src/main/java/oracle/kubernetes/operator/webhooks/model/DomainAdmissionReview.java
@@ -1,0 +1,69 @@
+// Copyright (c) 2022, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package oracle.kubernetes.operator.webhooks.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * AdmissionReview represents a request or a response object for an admission webhook call.
+ *
+ * <p></p>When a user performs an operation to a Kubernetes resource, Kubernetes ApiServer invokes all admission
+ * webhooks that have registered their interests in the resource. An AdmissionReview object contains the data that is
+ * communicated between the K8S ApiServer and a webhook. It is either a wrapper of
+ * an {@link AdmissionRequest} generated by the Kubernetes ApiServer and passed to a webhook, or a wrapper of an
+ * {@link AdmissionResponse} object generated by a webhook and sent back to the Kubernetes ApiServer.
+ * </p>
+ * <p>It implements a subset of the AdmissionReview interfaces described in <a href="https://kubernetes.io/docs/
+ reference/access-authn-authz/extensible-admission-controllers/#webhook-request-and-response">
+ Admission webhook request and response</a>.
+ * </p>
+ */
+public class DomainAdmissionReview extends AdmissionReview {
+  /**
+   * An {@link  AdmissionRequest}  that contains the details of the admission call.
+   */
+  @SerializedName("request")
+  @Expose
+  private DomainAdmissionRequest request;
+
+
+  public DomainAdmissionRequest getRequest() {
+    return request;
+  }
+
+  public void setRequest(DomainAdmissionRequest request) {
+    this.request = request;
+  }
+
+  public DomainAdmissionReview request(DomainAdmissionRequest request) {
+    setRequest(request);
+    return this;
+  }
+
+  public DomainAdmissionReview kind(String kind) {
+    setKind(kind);
+    return this;
+  }
+
+  public DomainAdmissionReview apiVersion(String apiVersion) {
+    setApiVersion(apiVersion);
+    return this;
+  }
+
+  public DomainAdmissionReview response(AdmissionResponse response) {
+    setResponse(response);
+    return this;
+  }
+
+  @Override
+  public String toString() {
+    return "AdmissionReview{"
+        + "apiVersion='" + apiVersion + '\''
+        + ", kind='" + kind + '\''
+        + ", request='" + request + '\''
+        + ", response='" + response + '\''
+        + '}';
+  }
+}
